[PaymentRequest] Move command providers to PaymentBundle
