+++ src/Sylius/Bundle/CoreBundle/DependencyInjection/SyliusCoreExtension.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\DependencyInjection;
 
-use Payum\Offline\OfflineGatewayFactory;
 use Sylius\Bundle\CoreBundle\Attribute\AsCatalogPromotionApplicatorCriteria;
 use Sylius\Bundle\CoreBundle\Attribute\AsCatalogPromotionPriceCalculator;
 use Sylius\Bundle\CoreBundle\Attribute\AsEntityObserver;
@@ -106,7 +105,6 @@ final class SyliusCoreExtension extends AbstractResourceExtension implements Pre
         $this->prependDoctrineMigrations($container);
         $this->prependJmsSerializerIfAdminApiBundleIsNotPresent($container);
         $this->prependSyliusOrderBundle($container, $config);
-        $this->prependPayum($container);
     }
 
     protected function getMigrationsNamespace(): string
@@ -184,15 +182,6 @@ final class SyliusCoreExtension extends AbstractResourceExtension implements Pre
         $container->prependExtensionConfig('sylius_order', [
             'autoconfigure_with_attributes' => $config['autoconfigure_with_attributes'] ?? false,
         ]);
-    }
-
-    private function prependPayum(ContainerBuilder $container): void
-    {
-        $loader = new XmlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
-
-        if (class_exists(OfflineGatewayFactory::class)) {
-            $loader->load('services/integrations/payum/offline.xml');
-        }
     }
 
     private function registerAutoconfiguration(ContainerBuilder $container): void
