--- src/Sylius/Bundle/CoreBundle/Resources/config/services/payment_request/offline/command_providers.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Sylius Sp. z o.o.
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<container xmlns="http://symfony.com/schema/dic/services"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
->
-    <services>
-        <defaults public="true" />
-
-        <service id="sylius.payment_request.command_provider.offline" class="Sylius\Bundle\CoreBundle\PaymentRequest\CommandProvider\ActionsCommandProvider">
-            <argument type="tagged_locator" tag="sylius.payment_request.command_provider.offline" index-by="action" />
-            <tag name="sylius.payment_request.command_provider" gateway-factory="offline" />
-        </service>
-
-        <service id="sylius.payment_request.command_provider.offline.capture" class="Sylius\Bundle\CoreBundle\PaymentRequest\CommandProvider\Offline\CapturePaymentRequestCommandProvider">
-            <tag name="sylius.payment_request.command_provider.offline" action="capture" />
-        </service>
-    </services>
-</container>
