+++ src/Sylius/Bundle/PaymentBundle/CommandProvider/Offline/CapturePaymentRequestCommandProvider.php
@@ -11,10 +11,10 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\CoreBundle\PaymentRequest\CommandProvider\Offline;
+namespace Sylius\Bundle\PaymentBundle\CommandProvider\Offline;
 
 use Sylius\Bundle\CoreBundle\PaymentRequest\Command\Offline\CapturePaymentRequest;
-use Sylius\Bundle\CoreBundle\PaymentRequest\CommandProvider\PaymentRequestCommandProviderInterface;
+use Sylius\Bundle\PaymentBundle\CommandProvider\PaymentRequestCommandProviderInterface;
 use Sylius\Component\Payment\Model\PaymentRequestInterface;
 
 final class CapturePaymentRequestCommandProvider implements PaymentRequestCommandProviderInterface
